Add code to initialize layers for the project
If the image path is provided, the image is read and the layers are
initialized accordingly (canvas size same as that of the image).

If the shape is provided, canvas size is equal to the shape provided else
the canvas size is equal to the default canvas shape as mentioned in
macros.py.