Refactor code
Earlier user was required to press a key on the main canvas window to
choose a tool/operation to perform and then the user was mostly required
to enter the details for the operation on the terminal which was a
untidy and confusing task.

Now the user does not need to press anything on the main canvas window.
The operation to perform is aasked on the terminal itself. The user
simply needs to enter the corresponding number in the terminal for a
operation and that operation is activated like earlier.