Fix bugs in rectangular marquee tool
Earlier, if the rectangle is selected and then moved, it was not
allowed to move outside the frame at all. This condition is
removed and now the rectangle can be moved completely
outside the frame also.

Masking out the part of the image that was not in the selected
region but was present in the selected region's bounding box
is also done. It will be useful in case of non- rectangular selected
regions.

Also, extra region from the final selected image is removed that
did not lie on any of the layer images that were to be copied.