Add code for quick selection tool
The code is not yet complete. Only selecting a foreground part is
complete. Deselecting the selected part using right mouse button
is not working yet.

Also, the code is not working in real time. It is slower for big
images, It is also obsereved as more and more regions are selected,
the code becomes even slower.